Move images to assets folder; remove CRA logo
