+++ frontend/src/assets/arrow_down.png
