+++ frontend/src/assets/home-icon.png
